I now see why the other state was being overwritten.
sending data to any of the actioncable subscriptions also has a echo to
each connection which returns the data, although it seems the second
time not all of the data is sent, which is confusing. it might be better
not to use the destructuring argument function signature as a result.